Fix size of a raw pointers
Issue: https://gitcode.com/openharmony/arkcompiler_ets_frontend/issues/8914

+++ ets2panda/util/eheap.h
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2025 Huawei Device Co., Ltd.
+ * Copyright (c) 2025-2026 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -128,8 +128,8 @@ public:
             if (raw == 0) {
                 return nullptr;
             }
-            auto ptr =
-                reinterpret_cast<T *>(ToUintPtr(gEHeapSpace->BaseAddr()) + (raw << EHeapSpace::ALLOC_LOG_ALIGNMENT));
+            auto ptr = reinterpret_cast<T *>(ToUintPtr(gEHeapSpace->BaseAddr()) +
+                                             (static_cast<uintptr_t>(raw) << EHeapSpace::ALLOC_LOG_ALIGNMENT));
             gEHeapSpace->AssertInRange(ptr);
             return ptr;
         }
